first snat and dnat implementation (#8)
currently snat and dnat are not really production ready.
map lookups are missing entirely, but the base logic 
is implemented and working. we need to figure out 
how to test these. this will be part of a separate MR.

also I added a bunch of includes. they are copied from
Cilium, Linux and libbpf. this is to enable easily
reproducible builds.