bpf map support for snat and dnat (#9)
dnat and snat targets can now be added dynamically using 
bpf maps. this commit implement the first initial version of
it. 

also implement tests to ensure that tcx links will be attached with
the correct configuration.

make func tests runnable on wsl (which means not being root). 
we will see how this works out in CI.