add linting ci job (#23)
for this to work correctly we have to include files that are
going to be embedded using go:embed. this is a problem 
in the case of docker image tar files used by tests in the 
image package, because they are very large. to work 
around this we have a pre-commit hook in place to
create empty files instead.