platformd: generate unique workload id (#25)
this commit generates a uuidv7 id for each created workload.

use-case for this id is basically to be able to group resources
of different origin under one single identifier. for example we need
this relation when choosing stat prefixes for envoy listeners. in this
case knowing which listener relates to which workload comes in handy
when debugging.