platformd: batch #1
this batch of changes fixes various things.

rename this to management-listen-sock, because this socket
will not not only provide proxy services. for now it will
serve all management-related apis.

the previous behavior would only apply the configuration
to nodes that where already known to the cache. this means
that the very first config update call, which initally
provides global resources, would not be put into the cache,
due to GetStatusKeys() not containing any node ids.

- pass image to UserSpecifiedImage in CreateContainer call
- do not crash when getsockopt bpf program is already pinned
- lb policy of original dst cluster must be CLUSTER_PROVIDED